Images2Thumbnails/requirements.txt
azure-storage-blob==12.8.1
azure-functions==1.7.2
imageio==2.13.3
Pillow==8.4.0